% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supraAdjacencyMatrix.R
\name{add_interlayer_links}
\alias{add_interlayer_links}
\title{Generate Node aligned Interlayer Edges for a Multilayer Network from the function \code{convert_to_intra_format}}
\usage{
add_interlayer_links(intra_data, interlayer_weight = 1)
}
\arguments{
\item{intra_data}{A list with \verb{$vertices} (node info) and \verb{$intra} (intralayer edges).}

\item{interlayer_weight}{Numeric, the weight for interlayer links (default: 1).}
}
\value{
A dataframe with interlayer edges.
}
\description{
Generate Node aligned Interlayer Edges for a Multilayer Network from the function \code{convert_to_intra_format}
}
