% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInteractionIntensity.r
\name{calc_interaction_intensity}
\alias{calc_interaction_intensity}
\title{Calcultates the interaction intensity of a food web using the metabolic theory and
the interaction dimensionality}
\usage{
calc_interaction_intensity(da, res_mm, con_mm, int_dim)
}
\arguments{
\item{da}{data.frame with the interactions body mass and type of interaction dimensionality}

\item{res_mm}{name of the column with the resource mass mean}

\item{con_mm}{name of the column with the consumer mass mean}

\item{int_dim}{name of the column with the interaction dimensionality}
}
\value{
a data.frame
}
\description{
It uses all the coeficients from \link{1}, taking into account the interaction dimensionality and body mass units in Kg.
For detritus or sediment the resource mass mean \code{res_mm} is generally < 0 thus resource body mass is calculated using
the equation S9 and supplementary figures 2c & d (kg) of the paper.
The values for the resource density $Xr$ where estimated according the equation S18 and supplementary figures 2i & j (individuals/m2 - m3)
}
\references{
\enumerate{
\item Pawar, S., Dell, A. I., & Van M. Savage. (2012). Dimensionality of consumer search space drives trophic interaction strengths. Nature, 486, 485. https://doi.org/10.1038/nature11131
}
}
