% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExtinctionIndices.r
\name{calc_QSS_extinction_dif}
\alias{calc_QSS_extinction_dif}
\title{Calculates the QSS difference between the full network and the network minus
one species}
\usage{
calc_QSS_extinction_dif(g, sp_list, nsim = 1000, ncores = 4, istrength = FALSE)
}
\arguments{
\item{g}{igraph network}

\item{sp_list}{list with the species/nodes we will delete for the comparison}

\item{nsim}{number of simulations to calculate QSS}

\item{ncores}{number of cores used to perform the operation}

\item{istrength}{if TRUE takes the weigth attribute of the network as interaction strength to
calculate QSS.}
}
\value{
a data.frame with:
\itemize{
\item the node deleted
\item the Anderson-Darling p-value of the comparison with the complete network
\item Kolmogorov-Smirnov p-value of the comparison with the complete network
\item If \code{istrength == TRUE} the Anderson-Darling p-value of the comparison with the null network
\item If \code{istrength == TRUE} Kolmogorov-Smirnov p-value of the comparison with the null network
\item median of QSS of the complete network
\item median of QSS of the network with the deleted node
\item difference between the two previous median QSS
\item If \code{istrength == TRUE} median of QSS of the null network
\item If \code{istrength == TRUE} difference between the median QSS of the network with the deleted node
and the null network.
}
}
\description{
The QSS determines the maximum eingenvalue of the community matrix (Jacobian) and characterizes
the linear stability of the network. This uses the function \code{\link[=calc_QSS]{calc_QSS()}} so it can take
into account the interaction strength if weights are present. The comparison is made using the Anderson-Darling test with
the function \code{\link[kSamples:ad.test]{kSamples::ad.test()}} and the Kolmogorov-Smirnov test \code{\link[stats:ks.test]{stats::ks.test()}}, both the p-values are reported as a
measure of strength of the diference. If istrength is TRUE it makes a comparison to a null model with the same species and links than the
reduced network but with all interaction strengths equal to the mean interacion strength.
}
\examples{
\dontrun{
g <- netData[[1]]

# Generate random weights
#
V(g)$weight <-  runif(vcount(g))

# Without interaction strength
#
calc_QSS_extinction_dif(g,V(g)$name[1:3],nsim=10,istrength = FALSE)

# With interaction strength
#
calc_QSS_extinction_dif(g,V(g)$name[1:3],nsim=10,istrength = TRUE)
}
}
