% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExtinctionIndices.r
\name{calc_QSS_extinction_dif_grp}
\alias{calc_QSS_extinction_dif_grp}
\title{Calculates the QSS difference between the full network and the network minus
a group of species}
\usage{
calc_QSS_extinction_dif_grp(
  g,
  sp_list,
  nsim = 1000,
  ncores = 4,
  istrength = FALSE
)
}
\arguments{
\item{g}{igraph network}

\item{sp_list}{list with the group pf species/nodes we will delete for the comparison}

\item{nsim}{number of simulations to calculate QSS}

\item{ncores}{number of cores used to perform the operation}

\item{istrength}{if TRUE takes the weight attribute of the network as interaction strength to
calculate QSS.}
}
\value{
a data.frame with:
\itemize{
\item Size of the network with deleted nodes
\item Connectance of the network with deleted nodes
\item Number of components of the network with deleted nodes
\itemize{
\item QSS of the complete network
\item QSS of the network with the deleted nodes
\item difference between the two previous QSS
}
}
}
\description{
The Quasi-sign stability is estimated with the maximum eigenvalue of the community matrix (Jacobian) and characterizes
the linear stability of the network. This uses the function \code{\link[=calc_QSS]{calc_QSS()}} so it can take
into account the interaction strength if weights are present. The comparison is made using the Anderson-Darling test with
the function \code{\link[kSamples:ad.test]{kSamples::ad.test()}} and the Kolmogorov-Smirnov test \code{\link[stats:ks.test]{stats::ks.test()}}, both the p-values are reported as a
measure of strength of the difference.
}
\examples{
\dontrun{
g <- netData[[1]]

# Generate random weights
#
V(g)$weight <-  runif(vcount(g))

# Without interaction strength
#
calc_QSS_extinction_dif_grp(g,V(g)$name[1:3],nsim=10,istrength = FALSE)

# With interaction strength
#
calc_QSS_extinction_dif_grp(g,V(g)$name[1:3],nsim=10,istrength = TRUE)
}
}
