% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExtinctionIndices.r
\name{calc_QSS_extinctions_seq}
\alias{calc_QSS_extinctions_seq}
\title{Calculates the QSS for an extinction sequence}
\usage{
calc_QSS_extinctions_seq(g_del, seq, nsim = 100, ncores = 0, istrength = FALSE)
}
\arguments{
\item{g_del}{igraph object for the deletion sequence}

\item{seq}{vector or list with the nodes for the extinction sequence, the
extinctions are incremental.}

\item{nsim}{number of simulations used in QSS calculation}

\item{ncores}{number of cores used to calculate QSS}

\item{istrength}{parameter istrength for QSS function}
}
\value{
data frame with the number of remaining nodes, the connectance, the number unconnected of components, the median QSS, and the name
of the last deleted node. The first row are the values for the complete network.
}
\description{
This functions calculates the QSS \code{\link[=calc_QSS]{calc_QSS()}} for a sequence of incremental deletions of species (nodes)
given by the vector \code{seq}, it does not produce secondary extinctions
}
\examples{
\dontrun{
g <- netData[[1]]

# Generate random weights
#
V(g)$weight <-  runif(vcount(g))

# Without interaction strength
#
calc_QSS_extinctions_seq(g,V(g)$name[1:3],nsim=10,istrength = FALSE)

# With interaction strength
#
calc_QSS_extinctions_seq(g,V(g)$name[1:3],nsim=10,istrength = TRUE)
}
}
