% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFunctionalIndices.r
\name{calc_quantitative_connectance}
\alias{calc_quantitative_connectance}
\alias{calcQuantitativeConnectance}
\title{Calculates the quantitative connectance, and the effective number of flows using the interaction matrix and a vector of species abundances}
\usage{
calc_quantitative_connectance(interM, d)
}
\arguments{
\item{interM}{per capita interaction strength matrix}

\item{d}{species' abundances vector}
}
\value{
A list with Cq,the quantitative connectance index and LDq, =
}
\description{
The Quantitative connectance (Cq) takes into account both the distribution of per capita interaction strengths
among species in the web and the distribution of species’ abundances and quantifies the diversity of network fluxes,
if all the species have the same flux is equal to the directed connectance.
The mean or effective number of flows impinging upon or emanating from a tipical node (LDq) is based the average flow diversity, and when all flows are equal is similar to linkage density. Both measures are based in Shannon information theory.
The total interaction flux is measured
as \code{T[i,j] <- d[i] * d[j] * interM[i,j]}. The effective Cq is calculated following the formulas in appendix 2 of \link{1}, LDq follows \link{2}
}
\examples{

# 3 predators 2 preys unequal fluxes
#
m <- matrix()
matrix(0,nrow=4,ncol=4)
m[1,2] <- m[1,3] <- m[3,4]<- .2
m[2,1] <- m[3,1] <- m[4,3] <- -2

calc_quantitative_connectance(m, c(1,1,1,1))

# Equal input and output fluxes

m <- matrix(0,nrow=4,ncol=4)
m[1,2] <- m[1,3] <- m[3,4]<- 2
m[2,1] <- m[3,1] <- m[4,3] <- -2
calc_quantitative_connectance(m, c(1,1,1,1))
}
\references{
\enumerate{
\item Fahimipour, A.K. & Hein, A.M. (2014). The dynamics of assembling food webs. Ecol. Lett., 17, 606–613
\item Ulanowicz, R.E. & Wolff, W.F. (1991). Ecosystem flow networks: Loaded dice? Math. Biosci., 103, 45–68
}
}
