% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdcentrality.R
\name{calc_svd_entropy}
\alias{calc_svd_entropy}
\title{Singular Value Decomposition (SVD) Entropy and Rank Calculation}
\usage{
calc_svd_entropy(A, threshold_factor = 1e-06)
}
\arguments{
\item{A}{An adjacency matrix (square, weighted or unweighted) or an igraph object.}

\item{threshold_factor}{A small value (default = 1e-6) to determine numerical rank approximation.}
}
\value{
A list containing:
\item{Rank}{Effective rank of the matrix.}
\item{Entropy}{SVD entropy of the network.}
}
\description{
This function calculates the entropy and rank of a network using Singular Value Decomposition (SVD).
It accepts either an adjacency matrix or an igraph object and processes it accordingly.
}
