% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calc_swness_zscore}
\alias{calc_swness_zscore}
\title{Calculation Small-world-ness z-scores}
\usage{
calc_swness_zscore(g, nullDist, sLevel = 0.01, ncores = 0,
  weights = NA)
}
\arguments{
\item{g}{igraph object}

\item{nullDist}{list of igraph object with the null model simulations}

\item{sLevel}{significance level to calculate CI (two tails)}

\item{ncores}{number of cores to use paralell computation, if 0 sequential processing is used.}
}
\value{
a list with two data frames: one with indices z-scores and CI

\item{Clustering}{ Clustering coefficient, measures the average fraction of pairs of neighbors of a node that are also neighbors of each other}
\item{PathLength}{ Mean of the shortest paths between all pair of vertices }
\item{Modularity}{ modularity measures how separated are different groups from each other, the algorithm \code{cluster_spinglass} was used to obtain the groups}
\item{zCC,zCP,zMO}{Z-scores of Clustering,PathLength and Modularity with respect to a random Erdos-Renyi null model}
\item{CClow,CChigh,CPlow,CPhigh,MOlow,MOhigh}{sLevel confidence intervals}
\item{SWness,SWnessCI}{ Small-world-ness and it CI value}
\item{isSW,isSWness}{ Logical variable signalling if the network is Small-world by the method of Marina 2018 or the method of Humprhies & Gurney 2008 }

Another data.frame with the values calculated for the nullDist.
}
\description{
The function calculates small-world-ness z-scores and 99\% CI intervals,
using as null model the list of networks in the nullDist parameter.
}
\examples{
\dontrun{
nullg <- generateERbasal(netData[[1]],10)
calc_swness_zscore(netData[[1]],nullg)
}

}
\references{
Marina, T. I., Saravia, L. A., Cordone, G., Salinas, V., Doyle, S. R., & Momo, F. R. (2018). Architecture of marine food webs: To be or not be a ‘small-world.’ PLoS ONE, 13(5), 1–13. https://doi.org/10.1371/journal.pone.0198217
}
