% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calc_topological_roles}
\alias{calc_topological_roles}
\title{Calc topological roles among network communities/modules}
\usage{
calc_topological_roles(g, nsim = 1000, ncores = 0)
}
\arguments{
\item{g}{an Igraph object with the network}

\item{nsim}{number of simulations with different community}
}
\value{
a  data frame with two numeric fields: within_module_degree, among_module_conn
}
\description{
Topological roles are characterized by two parameters: the standardized within-module degree \eqn{dz} and the among-module
connectivity participation coefficient \eqn{PC}.  The within-module degree is a z-score that measures how well a species is
connected to other species within its own module compared with a random graph. The participation coefficient \eqn{PC}
estimates the distribution of the links of species among modules
}
\examples{
#' \dontrun{

g <- netData[[2]]

tp <- calc_topological_roles(g,nsim=10)

}
}
\references{
\enumerate{
\item Guimerà, R. & Nunes Amaral, L.A. (2005). Functional cartography of complex metabolic networks. Nature, 433, 895–900
\item Saravia, L.A., Marina, T.I., De Troch, M. & Momo, F.R. (2018). Ecological Network assembly: how the regional meta web influence local food webs. bioRxiv, 340430, doi: https://doi.org/10.1101/340430
}
}
