% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFunctionalIndices.r
\name{calc_weighted_topological_indices}
\alias{calc_weighted_topological_indices}
\title{Function to calculate weighted network indices}
\usage{
calc_weighted_topological_indices(ig, ncores = 0)
}
\arguments{
\item{ig}{An igraph object with the weight attribute representing fluxes}

\item{ncores}{number of cores used to compute in parallel, if 0 sequential processing is used.}
}
\value{
a data.frame with the following fields:

\item{LD:}{ linkage density}
\item{Connectance:}{ directed Connectance}
\item{TLmean:}{ mean trophic level}
\item{TLmax:}{ maximum trophic level}
\item{LOmnivory:} { Level of omnivory, quantiﬁes mean of the variety in trophic levels of the preys of a consumer}
\item{Vulnerability:}{ mean of number of consumers per prey}
\item{VulSD:}{ the standard deviation of normalized Vulnerability}
\item{Generality:}{ mean number of prey per consumer}
\item{GenSD:}{ the standard deviation of normalized Generality}
}
\description{
The function calculates: weighted linkage density, connectance, generality and vulnerability and the SDs of the last two based on \link{1}
and the level of omnivory, mean and maximum trophic level based on \link{2} using \code{\link[NetIndices:TrophInd]{NetIndices::TrophInd()}} function.
The igraph networks must have the weight attribute.
}
\examples{

# Generate a test network

g <-   graph_from_literal( 1 -+ 4 -+ 7,2 -+ 5 -+7, 3-+6-+7, 7-+7, 4+-3, simplify = FALSE)

# Add weight

E(g)$weight <- sample(c(.1,.2,.8,.9),gsize(g),replace=TRUE)

calc_weighted_topological_indices(g)
}
\references{
\enumerate{
\item Bersier, LF. et al. (2002). Quantitative Descriptors Of Food-web Matrices. Ecology, 83(9), 2394–2407.
\item Kones, JK. et al. (2009). Are network indices robust indicators of food web functioning? A Monte Carlo approach.
Ecological Modelling, 220, 370–382.
}
}
