% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{calcIncoherence}
\alias{calcIncoherence}
\title{Calculate the incoherence index of a food web}
\usage{
calcIncoherence(g, ti = NULL)
}
\arguments{
\item{g}{an igraph object}

\item{ti}{trophic level vector if not supplied is calculated internally}
}
\value{
a data.frame with the following fields

      \item{Q}{incoherence (0=coherent)}
     \item{rQ}{ratio of Q with expected Q under null expectation of a random network given N=nodes L=links B=basal nodes Lb=basal links}
     \item{mTI}{mean trophic level}
     \item{rTI}{ratio of mTI with expected TI under the same null model expectation}
}
\description{
The incoherence index is based in how the species fit in discrete trophic levels
when Q is closer to 0 more coherent and stable is a food web.
}
\details{
Based on:
}
\examples{

calcIncoherence(netData[[1]])


}
\references{
Johnson, S., Domínguez-García, V., Donetti, L., & Muñoz, M. A. (2014). Trophic coherence determines food-web stability. Proceedings of the National Academy of Sciences , 111(50), 17923–17928. https://doi.org/10.1073/pnas.1409077111

Johnson, S., & Jones, N. S. (2017). Looplessness in networks is linked to trophic coherence. Proceedings of the National Academy of Sciences, 114(22), 5618–5623. https://doi.org/10.1073/pnas.1613786114
}
