% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calcModularitySWnessZScore}
\alias{calcModularitySWnessZScore}
\title{Calculation of Modularity and Small-world-ness z-scores}
\usage{
calcModularitySWnessZScore(g, nsim = 1000, sLevel = 0.01, paralell = TRUE)
}
\arguments{
\item{g}{igraph object}

\item{nsim}{number of simulations of the Erdos-Renyi random networks}

\item{sLevel}{significance level to calculate CI (two tails)}
}
\value{
a data frame with indices z-scores and CI
}
\description{
The function calculates modularity, number of groups and small-world-ness z-scores and 99\% CI intervals
using as null model Erdos-Renyi random networks with the same number of links and species.
}
\examples{

calcModularitySWnessZScore(netData[[1]])
}
\references{
Marina, T. I., Saravia, L. A., Cordone, G., Salinas, V., Doyle, S. R., & Momo, F. R. (2018). Architecture of marine food webs: To be or not be a ‘small-world.’ PLoS ONE, 13(5), 1–13. https://doi.org/10.1371/journal.pone.0198217
}
