% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFunctionalIndices.r
\name{calcQuantitativeConnectance}
\alias{calcQuantitativeConnectance}
\title{See \code{\link[=calc_quantitative_connectance]{calc_quantitative_connectance()}}}
\usage{
calcQuantitativeConnectance(interM, d)
}
\description{
See \code{\link[=calc_quantitative_connectance]{calc_quantitative_connectance()}}
}
