% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calcTopologicalIndices}
\alias{calcTopologicalIndices}
\title{Calculate topological indices for ecological networks}
\usage{
calcTopologicalIndices(ig, ncores = 0)
}
\arguments{
\item{ig}{vector of igraph objects}

\item{ncores}{number of cores used to compute in parallel, if 0 sequential processing is used.}
}
\value{
a data.frame with the following fields:

\item{Size:}{Number of species}
\item{Top:}{Number of top predator species}
\item{Basal:}{Number of basal especies}
\item{Omnivory:}{ Number of omnivory species, is the number }
\item{Links:}{ number of interactions}
\item{LD:}{ linkage density}
\item{Connectance:}{ Connectance}
\item{PathLength:}{ average path length}
\item{Clustering:}{ clustering coeficient}
\item{Cannib:}{ number of cannibalistic species}
\item{TLmean:}{ mean trophic level}
\item{TLmax:}{ maximum trophic level}
\item{Components:}{ number of weakly connected components}
}
\description{
Calculate topological indices for ecological networks
}
\examples{

topologicalIndices(netData)

# Generate a test network

g <-   graph_from_literal( 1 -+ 4 -+ 7,2 -+ 5 -+7, 3-+6-+7, 7-+7, 4+-3, simplify = FALSE)

topologicalIndices(g)
}
