% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{classify_topological_roles}
\alias{classify_topological_roles}
\title{Classify and plot topological roles}
\usage{
classify_topological_roles(tRoles, g, community = NULL)
}
\arguments{
\item{tRoles}{A data frame of calculated topological roles obtained from \code{\link[=calc_topological_roles]{calc_topological_roles()}}.}

\item{g}{An igraph network object.}

\item{community}{An igraph community object used to determine module membership. If NULL, the function calculates community structure using \code{cluster_spinglass()}.}
}
\value{
A list containing:
\itemize{
\item \code{hub_conn}: A data frame with node classifications and topological role values.
\item \code{gp}: A ggplot object visualizing the topological roles.
}
}
\description{
This function classifies species into four topological roles based on their within-module degree (dz) and among-module connectivity (PC), following the approach of Kortsch (2015). The classification is as follows:
\itemize{
\item \strong{modhub}: Species with high within-module degree (dz >= 2.5) and low among-module connectivity (PC < 0.625).
\item \strong{hubcon}: Species with high within-module degree (dz >= 2.5) and high among-module connectivity (PC >= 0.625).
\item \strong{modspe}: Species with low within-module degree (dz < 2.5) and low among-module connectivity (PC < 0.625).
\item \strong{modcon}: Species with low within-module degree (dz < 2.5) and high among-module connectivity (PC >= 0.625).
}
}
\details{
The function also generates a ggplot visualization of the species' topological roles, with nodes colored according to their module membership.
}
\examples{
\dontrun{

g <- netData[[2]]

tp <- calc_topological_roles(g,nsim=10)

classify_topological_roles(tp,g)
}
}
\references{
\enumerate{
\item Kortsch, S., Primicerio, R., Fossheim, M., Dolgov, A. V., & Aschan, M. (2015). Climate change alters the structure of arctic marine food webs due to poleward shifts of boreal generalists.
Proceedings of the Royal Society B: Biological Sciences, 282(1814), 20151546. https://doi.org/10.1098/rspb.2015.1546
\item Saravia, L. A., Marina, T. I., Kristensen, N. P., De Troch, M., & Momo, F. R. (2022). Ecological network assembly: How the regional metaweb influences local food webs.
Journal of Animal Ecology, 91(3), 630–642. https://doi.org/10.1111/1365-2656.13652
}
}
