% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supraAdjacencyMatrix.R
\name{convert_to_supra_adjacency}
\alias{convert_to_supra_adjacency}
\title{Convert Multilayer Network to Supra-Adjacency Matrix (SAM)}
\usage{
convert_to_supra_adjacency(
  igraph_list,
  layer_names = NULL,
  interlayer_weight = 1,
  isDirected = TRUE,
  sparse = TRUE,
  use_names = FALSE,
  interlayer = TRUE,
  clean = TRUE
)
}
\arguments{
\item{igraph_list}{A list of \code{igraph} objects (each representing a layer).}

\item{layer_names}{(Optional) A vector of names for each layer.}

\item{interlayer_weight}{Numeric, the weight for interlayer links (default: 1).}

\item{isDirected}{Logical, whether the network is directed.}

\item{sparse}{Logical, whether to return the supra-adjacency matrix as a sparse matrix (default: TRUE).}

\item{use_names}{Logical; if \code{TRUE}, edges will use node names instead of numeric IDs (default: \code{FALSE}).}

\item{interlayer}{Logical; if \code{FALSE}, no interlayer links are added (default: \code{TRUE}).}

\item{clean}{Logical; if \code{TRUE}, removes empty rows/columns (state nodes with no edges) from the output (default: TRUE).}
}
\value{
A list with:
\item{supra_matrix}{Supra-adjacency matrix with optional inter-layer links.}
\item{state_nodes_map}{Mapping of state nodes (layer-node combinations).}
}
\description{
Convert Multilayer Network to Supra-Adjacency Matrix (SAM)
}
