% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNetwork.r
\name{fromIgraphToMgraph}
\alias{fromIgraphToMgraph}
\title{From multiple igraph objects to 'mgraph'}
\usage{
fromIgraphToMgraph(g, types)
}
\arguments{
\item{types}{vector of types that represent the layers with the same length as mg}

\item{mg}{list of 'igraph' objects}
}
\value{
class 'mgraph' object
}
\description{
This functions takes a list of igraph objects and convert it to a 'mgraph' object.
}
\examples{

# Read a vector of files
#
\dontrun{

fileName <- c(system.file("extdata",  package = "EcoNetwork"))
dn <- list.files("inst/extdata",pattern = "^Kefi2015.*\\\\.txt$")
g <- readNetwork(dn,"inst/extdata", skipColumn = 2)
igraph2mgraph(g,c("Negative","Positive","Antagonistic"))

}
}
\seealso{
\code{\link[=readMultiplex]{readMultiplex()}}
}
