% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffling.R
\name{generate_shuffled_seq}
\alias{generate_shuffled_seq}
\title{Generate a Sequence of Shuffled Networks and Track Metric Evolution}
\usage{
generate_shuffled_seq(
  original_graph,
  max_delta = 10,
  delta = 10,
  directed = TRUE,
  weighted = TRUE,
  shuffle_func = shuffle_network_ws,
  modularity_func = cluster_infomap
)
}
\arguments{
\item{original_graph}{An igraph object representing the original network.}

\item{max_delta}{Integer. Number of total shuffling steps to perform.}

\item{delta}{Integer. Number of swaps per shuffling step.}

\item{directed}{Logical. Whether the network is directed (default: TRUE).}

\item{weighted}{Logical. Whether the network is weighted (default: TRUE).}

\item{shuffle_func}{Function. A network shuffling function (default: \code{shuffle_network_ws}).}

\item{modularity_func}{Function. A modularity calculation function (default: \code{cluster_infomap}).}
}
\value{
A list with:
\item{Networks}{A list of igraph objects representing the shuffled networks at each step.}
\item{Metrics}{A tibble containing the step number, SVD entropy, SVD Rank, and modularity score.}
}
\description{
This function generates a sequence of shuffled networks from an original graph, applying
incremental shuffling steps while tracking modularity, SVD rank and entropy .
}
