% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata for \code{netData}}
\format{
A data frame
\describe{
\item{Network}{Name of the network}
\item{Longitude}{Longitude of the geographical coordinates of the network}
\item{Latitude}{Latitude of the geographical coordinates of the network}
\item{Date}{Date range or approximate date when the data was collected}
\item{InteractionsSource}{Method used to determine the interactions}
\item{Reference}{Reference of the study from where data was obtained}
\item{ReferenceLink}{URL of the reference}
}
}
\source{
\url{https://doi.org/10.1371/journal.pone.0198217}
\enumerate{
\item Marina, T. I., Saravia, L. A., Cordone, G., Salinas, V., Doyle, S. R., & Momo, F. R. (2018). Architecture of marine food webs: To be or not be a ‘small-world.’ PLoS ONE, 13(5), 1–13. https://doi.org/10.1371/journal.pone.0198217
}
}
\usage{
metadata
}
\description{
A data.frame with extended information about the 29 networks from (1), included in \code{netData}
}
\keyword{datasets}
