% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{netData}
\alias{netData}
\title{Compilation of Ecological Networks (food webs) in igraph format.}
\format{
A list with 29 igraph network objects
}
\source{
\url{https://doi.org/10.1371/journal.pone.0198217}
}
\usage{
netData
}
\description{
A dataset 29 curated and highly-resolved marine networks (food webs), updated from (1)
}
\references{
\enumerate{
\item Marina, T. I. et al. 2018. Architecture of marine food webs: To be or not be a ‘small-world.’ - PLoS ONE 13: 1–13.
}
}
\keyword{datasets}
