% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{parTopologicalIndices}
\alias{parTopologicalIndices}
\title{Paralell version of topologicalIndices}
\usage{
parTopologicalIndices(ig)
}
\arguments{
\item{ig}{vector of igraph objects}
}
\value{
the same as topologicalIndices
}
\description{
Paralell version of topologicalIndices
}
\examples{

parTopologicalIndices(netData)
}
