% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.r
\name{plot_troph_level}
\alias{plot_troph_level}
\alias{plotTrophLevel}
\title{Plot ecological network organized by trophic level, with node size determined by the node degree, and modules.}
\usage{
plot_troph_level(
  g,
  vertexLabel = FALSE,
  vertexSizeFactor = 5,
  vertexSizeMin = 5,
  tk = FALSE,
  modules = FALSE,
  lMat = NULL,
  weights = NA,
  community_obj = NULL,
  bpal = NULL,
  maxTL = NULL,
  edge.width = NULL,
  ...
)
}
\arguments{
\item{vertexLabel}{logical plot vertex labels}

\item{vertexSizeFactor}{numeric factor to determine the size of the label with degree}

\item{vertexSizeMin}{numeric determine the minimum size of the label}

\item{tk}{TRUE generate an interactive plot using tkplot and returns a matrix with coordinates from \code{\link[igraph:tkplot]{igraph::tkplot.getcoords()}}}

\item{modules}{if TRUE plot modules in the x axis, obtained with the cluster spinglass algorithm}

\item{lMat}{Matrix of postions for the nodes}

\item{weights}{The weights of the edges for the \code{\link[igraph:cluster_spinglass]{igraph::cluster_spinglass()}} community detection, either a numeric vector, NULL, or NA.
if NULL and the network has a 'weight' attribute then that will be used,
if NA then the 'weight' attributte is not considered.}

\item{community_obj}{Insteado of calculating modules/communities with cluster spinglass take a community object}

\item{bpal}{if NULL it uses the "RdYlGn" RColorBrewer palette, else must be a vector of colors of length 11.}

\item{maxTL}{maximum trophic level to draw y-axis}

\item{edge.width}{if NULL edge width is fixed in 0.3, if numberic takes into account edge weights multiplied by this value to draw edge widths,
the parameter weights determine the behavior.}

\item{...}{Addittional parameters to the plot function}

\item{ig}{igraph object}
}
\value{
returns a plot and if tk==TRUE returns a layout matrix
}
\description{
Plot ecological network organized by trophic level, with node size determined by the node degree, and modules.
}
\examples{

plot_troph_level(netData[[1]])
}
