% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetworkGgplot.R
\name{plot_troph_level_ggplot}
\alias{plot_troph_level_ggplot}
\title{Plot an Ecological Network Organized by Trophic Level}
\usage{
plot_troph_level_ggplot(
  g,
  vertexSizeFactor = 5,
  vertexSizeMin = 5,
  modules = FALSE,
  weights = NA,
  community_obj = NULL,
  maxTL = NULL,
  use_numbers = FALSE,
  label_size = 4
)
}
\arguments{
\item{g}{An \code{igraph} object representing the ecological network.}

\item{vertexSizeFactor}{Numeric factor to determine node size based on degree (default: 5).}

\item{vertexSizeMin}{Numeric value for the minimum node size (default: 5).}

\item{modules}{Logical; if \code{TRUE}, nodes are grouped by community modules detected using \code{cluster_spinglass()} (default: \code{FALSE}).}

\item{weights}{A numeric vector, \code{NULL}, or \code{NA}, specifying edge weights for community detection.
If \code{NULL}, the function uses the \code{weight} attribute of \code{g} if available. If \code{NA}, weights are ignored (default: \code{NA}).}

\item{community_obj}{An optional community detection object. If \code{NULL}, the function calculates modules automatically (default: \code{NULL}).}

\item{maxTL}{Numeric, maximum trophic level to display on the y-axis. If \code{NULL}, it is determined automatically (default: \code{NULL}).}

\item{use_numbers}{Logical; if \code{TRUE}, nodes are labeled with numeric IDs instead of species names to reduce label overlap in large networks (default: \code{FALSE}).}

\item{label_size}{Numeric, font size for node labels (default: 4).}
}
\value{
A \code{ggplot} object visualizing the trophic structure of the network.
If \code{use_numbers = TRUE}, it also returns a \code{tibble} mapping numeric labels to species names.
}
\description{
This function visualizes a food web using \code{ggplot2}, where nodes represent species and edges represent interactions.
Nodes are positioned by trophic level on the y-axis, and optionally, community modules on the x-axis.
}
\examples{

g <- netData$Angola
plot_troph_level_ggplot(g, modules = TRUE, use_numbers = TRUE)


}
