% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetworkGgplot.R
\name{plot_troph_level_ggraph}
\alias{plot_troph_level_ggraph}
\title{Plot an Ecological Network Organized by Trophic Level using ggraph}
\usage{
plot_troph_level_ggraph(
  g,
  vertexSizeFactor = 5,
  vertexSizeMin = 5,
  modules = FALSE,
  weights = NA,
  community_obj = NULL,
  use_numbers = FALSE,
  label_size = 4,
  arrow_size = 0.1,
  node_weights = NA
)
}
\arguments{
\item{g}{An \code{igraph} network.}

\item{vertexSizeFactor}{Factor for node size scaling (default 5).}

\item{vertexSizeMin}{Minimum node size (default 5).}

\item{modules}{If TRUE, uses \code{cluster_spinglass} for community detection.}

\item{weights}{Edge weights for clustering (default NA).}

\item{community_obj}{Optional community detection object.}

\item{use_numbers}{Label nodes by numeric ID.}

\item{label_size}{Text size.}

\item{arrow_size}{Arrowhead size.}

\item{node_weights}{Optional numeric vector for node weights. If NA, uses degree.
If it is NULL and the input graph has a ‘weight’ vertex attribute, then that attribute will be used.}
}
\value{
A \code{ggraph} plot.
}
\description{
This function visualizes a food web using \code{ggplot2}, where nodes represent species and edges represent interactions.
Nodes are positioned by trophic level on the y-axis, and optionally, community modules on the x-axis.
Is similar to \code{\link[=plot_troph_level_ggplot]{plot_troph_level_ggplot()}}, but uses \code{ggraph} for layout and rendering, which make better arrows.
}
