% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plotEcoNetworkTrophLevel}
\alias{plotEcoNetworkTrophLevel}
\title{Plot ecological network organized by trophic level, with node size determined by the node degree}
\usage{
plotEcoNetworkTrophLevel(g, vertexLabel = FALSE, vertexSizeFactor = 5)
}
\arguments{
\item{vertexLabel}{logical plot vertex labels}

\item{vertexSizeFactor}{numeric factor to determine the size of the label with degree}

\item{ig}{igraph object}
}
\value{
a plot
}
\description{
Plot ecological network organized by trophic level, with node size determined by the node degree
}
\examples{

plotEcoNetworkTrophLevel(netData[[1]])
}
