% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.r
\name{plotTrophLevel}
\alias{plotTrophLevel}
\title{Plot ecological network organized by trophic level, with node size determined by the node degree, and modules}
\usage{
plotTrophLevel(g, vertexLabel = FALSE, vertexSizeFactor = 5,
  tk = FALSE, modules = FALSE, lMat = NULL, weights = NA, ...)
}
\arguments{
\item{vertexLabel}{logical plot vertex labels}

\item{vertexSizeFactor}{numeric factor to determine the size of the label with degree}

\item{tk}{TRUE generate an interactive plot using tkplot}

\item{modules}{if TRUE plot modules in the x axis, obtained with the cluster spinglass algorithm}

\item{lMat}{Matrix of postions for the nodes}

\item{weights}{edge attribute for the \code{\link[igraph:cluster_spinglass]{igraph::cluster_spinglass()}} community detection,
if NA weight attributte is not considered, NULL weight attribute is taken into account,
other edge attribute to take into account for weight.}

\item{...}{Addittional parameters to the plot function}

\item{ig}{igraph object}
}
\value{
if tk==TRUE returns a layout matrix, else returns a plot
}
\description{
Plot ecological network organized by trophic level, with node size determined by the node degree, and modules
}
\examples{

plotTrophLevel(netData[[1]])
}
