% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{readEcoNetwork}
\alias{readEcoNetwork}
\title{Read ecological networks in CSV format as edge list or adyacency matrix}
\usage{
readEcoNetwork(fileName)
}
\arguments{
\item{fileName}{Filename of the csv formated network}
}
\value{
an igraph object
}
\description{
Read ecological networks in CSV format as edge list or adyacency matrix
}
\examples{
readEcoNetwork("econetwork.csv")
}
