% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{topologicalIndicesEcoNetwork}
\alias{topologicalIndicesEcoNetwork}
\title{Calculate topological indices for ecological networks}
\usage{
topologicalIndicesEcoNetwork(ig)
}
\arguments{
\item{ig}{vector of igraph objects}
}
\value{
a data.frame with the following fields:
        Size: Number of species
        Top:  Number of top predator species
        Basal: Number of basal especies
        Links: number of interactions
        LD: linkage density
        Connectance: Connectance
        PathLength: average path length
        Clustering: clustering coeficient
        Cannib: number of cannibalistic species
}
\description{
Calculate topological indices for ecological networks
}
