% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calc_modularity}
\alias{calc_modularity}
\title{Calculation of Modularity for a list of igraph objects}
\usage{
calc_modularity(ig, ncores = 0)
}
\arguments{
\item{ig}{list of igraph objects to calculate modularity}

\item{ncores}{number of cores used to compute in parallel, if 0 sequential processing is used.}
}
\value{
a data.frame with the field Modularity
}
\description{
The function calculates modularity of the list of networks in the nullDist parameter.
Modularity is calculated using the \code{\link[igraph:cluster_spinglass]{igraph::cluster_spinglass()}}
if the parameter weights is NULL the atribute "weigths" is used, or if it has the name of an network attribute,
that is used as a weigth to build the modules, when this parameter is NA then no weigth is used.
Only works for one component networks.
}
\examples{
\dontrun{
nullg <- generateERbasal(netData[[1]],10)
calc_modularity(nullg)
}

}
