% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFunctionalIndices.r
\name{calc_QSS}
\alias{calc_QSS}
\title{Calc the Quasi Sign Stability measure for antagonistic (predator-prey) or mgraph networks with multiple interactions.}
\usage{
calc_QSS(
  ig,
  nsim = 1000,
  ncores = 0,
  negative = -10,
  positive = 0.1,
  selfDamping = -1,
  istrength = FALSE,
  returnRaw = FALSE
)
}
\arguments{
\item{ig}{igraph or a list of igraph networks or mgraph network}

\item{nsim}{number of simulations to calculate QSS, if the number of simulations is 1 then it calculates the maximum eingenvalue for the mean of interaction
strength, if \code{istrength==FALSE} the mean interaction strength are \code{negative/2}, \code{positive/2}, \code{selfDamping/2}.}

\item{ncores}{number of cores to use in parallel comutation if 0 it uses sequential processing}

\item{negative}{the maximum magnitude of the negative interaction (the effect of the predator on the prey) must be <= 0}

\item{positive}{the maximum magnitude of the positive interaction (the effect of the prey on the predator) must be >= 0}

\item{selfDamping}{the maximum magnitude of the self-limitation (the effect of the species on itself) must be <= 0, only for species with links to itself.}

\item{istrength}{If TRUE takes the weigth attribute of the network as interaction strength.}

\item{returnRaw}{if TRUE returns all the values of the maximum eingenvalues}
}
\value{
if parameter \code{returnRaw} is \code{FALSE} returns a data.frame with the QSS, and MEing, the mean of the real part of the maximum eingenvalue.
If \code{returnRaw} is \code{TRUE} it returns the values of randomized real part of maximum eingenvalue (maxre)
}
\description{
The proportion of matrices that are locally stable, these matrices are created by sampling the values of the community matrix
(the Jacobian) from a uniform distribution, preserving the sign structure \link{1}. If the 'ig' parameter is
an \code{mgraph} network it needs to have been built with the order \code{c("Competitive", "Mutualistic", "Trophic")}
It also calculates the mean of the real part of the maximum eingenvalue, which is also a measure of stability \link{2}.
It uses a uniform distribution between 0 and maximum values given by the parameters \code{negative}, \code{positive} and \code{selfDamping},
corresponding to the sign of interactions and self-limitation effect \link{3,4}.
If the edges of the networks have a weigth attribute and \code{istrength} parameter is true, weigth will be used as interaction strength,
then the limits of the uniform distribution will be \code{negative*-x}, \code{positive*x}, \code{selfDamping*x}, where x is the value of the weigth for the edge.
If the values of these parameters are 0 then there is no interaction of that kind. The default values for \code{negative}, \code{positive} and \code{selfDumping}
assume an ecological transfer efficience of 1\%.
}
\examples{
\dontrun{

g <- netData[[2]]

tp <- calc_QSS(g)

# Read Multiplex network and calculate QSS

fileName <- c(system.file("extdata",  package = "multiweb"))
dn <- list.files(fileName,pattern = "^Kefi2015.*\\\\.txt$")
gt <- readMultiplex(dn,types,"inst/extdata", skipColumn = 2)
calc_QSS(gt)

}
}
\references{
\enumerate{
\item Allesina, S. & Pascual, M. (2008). Network structure, predator - Prey modules, and stability in large food webs.
Theor. Ecol., 1, 55–64.
\item Grilli, J., Rogers, T. & Allesina, S. (2016). Modularity and stability in ecological communities. Nat. Commun., 7, 12031
\item Monteiro, A.B. & Del Bianco Faria, L. (2017). Causal relationships between population stability and food-web topology. Functional Ecology, 31, 1294–1300.
\item Borrelli, J. J. 2015. Selection against instability: stable subgraphs are most frequent in empirical food webs. - Oikos 124: 1583–1588.
}
}
