% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFunctionalIndices.r
\name{calc_QSS_extinction_dif}
\alias{calc_QSS_extinction_dif}
\title{Calculates the QSS difference between the full network and the network minus
one species}
\usage{
calc_QSS_extinction_dif(g, sp_list, nsim = 1000, ncores = 4, istrength = FALSE)
}
\arguments{
\item{g}{igraph network}

\item{sp_list}{list with the species/nodes we will delete for the comparison}

\item{nsim}{number of simulations to calculate QSS}

\item{ncores}{number of cores used to perform the operation}

\item{istrength}{if TRUE takes the weigth attribute of the network as interaction strength to
calculate QSS.}
}
\value{
a data.frame with: the node deleted, Anderson-Darling p-value, median QSS of the complete network, median QSS of the network with deleted
node, difference between them.
}
\description{
The QSS determines the maximum eingenvalue of the community matrix (Jacobian) and characterizes
the linear stability of the network. This uses the function \code{\link[=calc_QSS]{calc_QSS()}} so it can take
into account the interaction strength if weigths are present. The comparison is made using the Anderson-Darling test with
the function \code{\link[kSamples:ad.test]{kSamples::ad.test()}} and the p-value is reported as a measure of strength of the diference.
}
\examples{
\dontrun{
g <- netData[[1]]

# Generate random weights
#
V(g)$weight <-  runif(vcount(g))

# Without interaction strength
#
calc_QSS_extinction_dif(g,V(g)$name[1:3],nsim=10,istrength = FALSE)

# With interaction strength
#
calc_QSS_extinction_dif(g,V(g)$name[1:3],nsim=10,istrength = TRUE)
}
}
