% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calc_topological_indices}
\alias{calc_topological_indices}
\alias{calcTopologicalIndices}
\title{Calculate topological indices for ecological networks. The mean trophic level
and Omnivory and the Level of omnivory are calculated with the function \code{\link[NetIndices:TrophInd]{NetIndices::TrophInd()}}}
\usage{
calc_topological_indices(ig, ncores = 0)
}
\arguments{
\item{ig}{vector of igraph objects}

\item{ncores}{number of cores used to compute in parallel, if 0 sequential processing is used.}
}
\value{
a data.frame with the following fields:

\item{Size:}{Number of species}
\item{Top:}{Number of top predator species}
\item{Basal:}{Number of basal especies}
\item{Omnivory:}{ Proportion of omnivorous species }
\item{Links:}{ number of interactions}
\item{LD:}{ linkage density}
\item{Connectance:}{ directed Connectance}
\item{PathLength:}{ average path length}
\item{Clustering:}{ clustering coeficient}
\item{Cannib:}{ number of cannibalistic species}
\item{TLmean:}{ mean trophic level}
\item{TLmax:}{ maximum trophic level}
\item{LOmnivory:} { Level of omnivory, quantiﬁes mean of the variety in trophic levels of the preys of a consumer}
\item{Components:}{ number of weakly connected components}
\item{Vulnerability:}{ mean of number of consumers per prey}
\item{VulSD:}{ the standard deviation of normalized Vulnerability}
\item{Generality:}{ mean number of prey per consumer}
\item{GenSD:}{ the standard deviation of normalized Generality}
}
\description{
Calculate topological indices for ecological networks. The mean trophic level
and Omnivory and the Level of omnivory are calculated with the function \code{\link[NetIndices:TrophInd]{NetIndices::TrophInd()}}
}
\examples{

calc_topological_indices(netData)

# Generate a test network

g <-   graph_from_literal( 1 -+ 4 -+ 7,2 -+ 5 -+7, 3-+6-+7, 7-+7, 4+-3, simplify = FALSE)

calc_topological_indices(g)

}
