% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calc_topological_roles}
\alias{calc_topological_roles}
\title{Calculate Topological Roles Within Network Modules}
\usage{
calc_topological_roles(g, nsim = 1, ncores = 0, community = NULL)
}
\arguments{
\item{g}{An \code{igraph} object representing the network.}

\item{nsim}{Integer. Number of independent runs of the Spinglass community detection algorithm (default is 1).}

\item{ncores}{Integer. Number of cores to use for parallel computation. Set to 0 for sequential processing.}

\item{community}{Optional. A community detection object (class \code{communities}) from an alternative algorithm. If \code{NULL}, \code{\link[igraph]{cluster_spinglass}} is used.}
}
\value{
A data frame with the following columns:
\describe{
\item{node}{Node ID.}
\item{within_module_degree}{Within-module degree z-score (\eqn{z}).}
\item{among_module_conn}{Participation coefficient (\eqn{PC}).}
}
}
\description{
This function identifies the topological roles of nodes (e.g., species) in an ecological or interaction network based on their position within and among modules.
Modules (or communities) are detected using the \code{\link[igraph]{cluster_spinglass}} algorithm by default, unless an alternative community object is provided.
}
\details{
Topological roles are quantified using two parameters:
\itemize{
\item \eqn{z} (within-module degree z-score): quantifies how well-connected a node is to others within its own module, relative to other members of the same module.
\item \eqn{PC} (participation coefficient): quantifies how evenly a node’s links are distributed across different modules.
}

When using the Spinglass algorithm, which is stochastic, multiple runs can be performed by setting the \code{nsim} parameter to repeat the procedure.
}
\examples{
#' \dontrun{

g <- netData[[2]]

tp <- calc_topological_roles(g,nsim=10)

# using a community object

m <- cluster_walktrap(g)

tp <- calc_topological_roles(g,community=m)
}
}
\references{
\enumerate{
\item Guimerà, R. & Nunes Amaral, L.A. (2005). Functional cartography of complex metabolic networks. \emph{Nature}, 433, 895–900.
\item Kortsch, S., Primicerio, R., Fossheim, M., Dolgov, A.V., & Aschan, M. (2015). Climate change alters the structure of Arctic marine food webs due to poleward shifts of boreal generalists. \emph{Proceedings of the Royal Society B: Biological Sciences}, 282(1814), 20151546. \doi{10.1098/rspb.2015.1546}
}
}
