% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calcIncoherence}
\alias{calcIncoherence}
\title{Calculate the incoherence index of a food web}
\usage{
calcIncoherence(ig, ti = NULL, ncores = 0)
}
\arguments{
\item{ti}{trophic level vector if not supplied is calculated internally}

\item{ncores}{number of cores used to compute in parallel, if 0 sequential processing is used.}

\item{g}{an igraph object}
}
\value{
a data.frame with the following fields

\item{Q}{incoherence (0=coherent)}
\item{rQ}{ratio of Q with expected Q under null expectation of a random network given N=nodes L=links B=basal nodes Lb=basal links}
\item{mTI}{mean trophic level}
\item{rTI}{ratio of mTI with expected TI under the same null model expectation than Q}
}
\description{
The incoherence index is based in how the species fit in discrete trophic levels
when Q is closer to 0 more coherent and stable is a food web, and the less omnivory it has.
It calculates the trophic level using the package NetIndices, or optionally the trophic levels could be passed
by parameter as a vector but in that case that the same vector will be used for all the networks.
}
\details{
Based on:
}
\examples{

calcIncoherence(netData[[1]])


}
\references{
Johnson, S., Domínguez-García, V., Donetti, L., & Muñoz, M. A. (2014). Trophic coherence determines food-web stability. Proceedings of the National Academy of Sciences , 111(50), 17923–17928. https://doi.org/10.1073/pnas.1409077111

Johnson, S., & Jones, N. S. (2017). Looplessness in networks is linked to trophic coherence. Proceedings of the National Academy of Sciences, 114(22), 5618–5623. https://doi.org/10.1073/pnas.1613786114
}
