% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{calcModularitySWnessZScore}
\alias{calcModularitySWnessZScore}
\title{Calculation of Modularity and Small-world-ness z-scores}
\usage{
calcModularitySWnessZScore(g, nsim = 1000, sLevel = 0.01, ncores = 0)
}
\arguments{
\item{g}{igraph object}

\item{nsim}{number of simulations of the Erdos-Renyi random networks}

\item{sLevel}{significance level to calculate CI (two tails)}

\item{ncores}{number of cores to use paralell computation}
}
\value{
a data frame with indices z-scores and CI

\item{Clustering}{ Clustering coefficient, measures the average fraction of pairs of neighbors of a node that are also neighbors of each other}
\item{PathLength}{ Mean of the shortest paths between all pair of vertices }
\item{Modularity}{ modularity measures how separated are different groups from each other, the algorithm \code{cluster_spinglass} was used to obtain the groups}
\item{zCC,zCP,zMO}{Z-scores of Clustering,PathLength and Modularity with respect to a random Erdos-Renyi null model}
\item{CClow,CChigh,CPlow,CPhigh,MOlow,MOhigh}{sLevel confidence intervals}
\item{SWness,SWnessCI}{ Small-world-ness and it CI value}
\item{isSW,isSWness}{ Logical variable signalling if the network is Small-world by the method of Marina 2018 or the method of Humprhies & Gurney 2008 }
}
\description{
The function calculates modularity, number of groups and small-world-ness z-scores and 99\% CI intervals
using as null model Erdos-Renyi random networks with the same number of links and species.
}
\examples{

calcModularitySWnessZScore(netData[[1]])
}
\references{
Marina, T. I., Saravia, L. A., Cordone, G., Salinas, V., Doyle, S. R., & Momo, F. R. (2018). Architecture of marine food webs: To be or not be a ‘small-world.’ PLoS ONE, 13(5), 1–13. https://doi.org/10.1371/journal.pone.0198217
}
