% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r
\name{classify_topological_roles}
\alias{classify_topological_roles}
\title{Classify and plot topological roles}
\usage{
classify_topological_roles(tRoles, g, community = NULL, plt = FALSE)
}
\arguments{
\item{tRoles}{Calculated topological roles with the function \code{\link[=calc_topological_roles]{calc_topological_roles()}}}

\item{g}{Igraph network object}

\item{community}{Igraph community object used to calculate the topological roles}

\item{plt}{logical whether a plot is generated}
}
\value{
a data.frame with the classified topological roles
}
\description{
We use the topological roles calculated with \code{\link[=calc_topological_roles]{calc_topological_roles()}} and categorize them following the approach of
Kortsch (2015), that defines species roles based in two thresholds \eqn{PC=0.625} and \eqn{dz=2.5}.
If a species had at least 60\% of links within its own module then \eqn{PC<0.625}, and if it also had \eqn{dz\ge 2.5}, thus it was
classified as a module hub, labeled \strong{modhub}. If a species had \eqn{PC<0.625} and \eqn{dz<2.5}, then it was called a peripheral
or specialist, labeled \strong{modspe}. Species that had \eqn{PC\ge0.625} and \eqn{dz<2.5} were considered module connectors \strong{modcon}.
Finally, if a species had \eqn{PC\ge0.625} and \eqn{dz\ge 2.5}, then it was classified as a super-generalist or hub-connector \strong{hubcon}.
}
\examples{
\dontrun{

g <- netData[[2]]

tp <- calc_topological_roles(g,nsim=10)

classify_topological_roles(tp,g,plt=TRUE)
}
}
\references{
\enumerate{
\item Kortsch, S., Primicerio, R., Fossheim, M., Dolgov, A. V & Aschan, M. (2015). Climate change alters the structure of arctic marine food webs due to poleward shifts of boreal generalists. Proc. R. Soc. B Biol. Sci., 282
\item Saravia, L.A., Marina, T.I., De Troch, M. & Momo, F.R. (2018). Ecological Network assembly: how the regional meta web influence local food webs. bioRxiv, 340430, doi: https://doi.org/10.1101/340430
}
}
