% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.r
\name{curveBall}
\alias{curveBall}
\title{Curbe ball algorithm to generate random networks from a given network with the same colums and row totals}
\usage{
curveBall(g, nsim = 1000)
}
\arguments{
\item{g}{igraph object to extract adjacency matrix}

\item{nsim}{number of generated random networks}
}
\value{
a list of randomized igraph objects
}
\description{
Based on:
}
\examples{

curveBall(netData[[1]])
}
\references{
Strona, G. et al. 2014. A fast and unbiased procedure to randomize ecological binary matrices with
fixed row and column totals. -Nat. Comm. 5: 4114. doi: 10.1038/ncomms5114
}
