% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNetwork.r
\name{fromGLVadjToIgraph}
\alias{fromGLVadjToIgraph}
\title{From Generalized Lotka Volterra adjacency matrix to igraph object}
\usage{
fromGLVadjToIgraph(glvAdj, spc)
}
\arguments{
\item{glvAdj}{numeric matrix Generalized Lotka-Volterra adjacency matrix}

\item{spc}{numeric vector of present Species}
}
\value{
an igraph object
}
\description{
It counts predator-prey/Antagonistic interactions like 1 edge,
competition and mutualisms are counted as is: two edges or one edge
}
