% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{netData}
\alias{netData}
\title{Compilation of Ecological Networks (food webs) in igraph format.}
\format{
A list with 23 igraph network objects
}
\source{
\url{https://doi.org/10.1371/journal.pone.0198217}
}
\usage{
netData
}
\description{
A dataset 23 marine ecological networks (food webs) from \link{1}
}
\references{
\link{1} Marina, T. I. et al. 2018. Architecture of marine food webs: To be or not be a ‘small-world.’ - PLoS ONE 13: 1–13.
}
\keyword{datasets}
