% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopologicalIndices.r, R/data.r
\docType{data}
\name{netData}
\alias{netData}
\title{Set of 29 marine food webs}
\format{
a list of igraph objects

A list with 23 igraph network objects
}
\source{
\url{https://doi.org/10.1371/journal.pone.0198217}
\enumerate{
\item Marina, T. I., Saravia, L. A., Cordone, G., Salinas, V., Doyle, S. R., & Momo, F. R. (2018). Architecture of marine food webs: To be or not be a ‘small-world.’ PLoS ONE, 13(5), 1–13. https://doi.org/10.1371/journal.pone.0198217
}

\url{https://doi.org/10.1371/journal.pone.0198217}
}
\usage{
netData

netData
}
\description{
A named list of igraph objects containing 29 curated and highly-resolved marine networks, updated from \link{1}.

A dataset 23 marine ecological networks (food webs) from \link{1}
}
\references{
\link{1} Marina, T. I. et al. 2018. Architecture of marine food webs: To be or not be a ‘small-world.’ - PLoS ONE 13: 1–13.
}
\keyword{datasets}
