% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stability_threshold.R
\name{plot_stability_curve}
\alias{plot_stability_curve}
\title{Plot stability threshold search curve}
\usage{
plot_stability_curve(result)
}
\description{
Plot stability threshold search curve
}
