% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetworkGgplot.R
\name{plot_troph_level_ggplot}
\alias{plot_troph_level_ggplot}
\title{Plot an Ecological Network Organized by Trophic Level}
\usage{
plot_troph_level_ggplot(
  g,
  vertexSizeFactor = 5,
  vertexSizeMin = 5,
  modules = FALSE,
  weights = NA,
  node_weights = NA,
  community_obj = NULL,
  use_numbers = FALSE,
  label_size = 4,
  arrow_size = 0.15,
  shorten_factor = 0.005
)
}
\arguments{
\item{g}{An \code{igraph} object representing the ecological network.}

\item{vertexSizeFactor}{Numeric factor to scale node size (default: 5).}

\item{vertexSizeMin}{Minimum node size (default: 5).}

\item{modules}{Logical; if \code{TRUE}, nodes are grouped by community modules (default: \code{FALSE}).}

\item{weights}{Edge weights for community detection (default: \code{NA}).}

\item{node_weights}{Optional numeric vector for node weights. If \code{NULL} and the graph has a vertex attribute called \code{weight}, that attribute is used. If not, node degree is used as fallback.}

\item{community_obj}{Optional community detection object.}

\item{use_numbers}{Logical; if \code{TRUE}, label nodes by numeric ID (default: \code{FALSE}).}

\item{label_size}{Label font size (default: 4).}

\item{arrow_size}{Arrowhead size (default: 0.15).}

\item{shorten_factor}{Factor to shorten edges for arrows (default: 0.005).}
}
\value{
A \code{ggplot} object visualizing the trophic structure of the network.
}
\description{
This function visualizes a food web using \code{ggplot2}, where nodes represent species and edges represent interactions.
Nodes are positioned by trophic level on the y-axis, and optionally, community modules on the x-axis.
}
