% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plotTrophLevel}
\alias{plotTrophLevel}
\title{Plot ecological network organized by trophic level, with node size determined by the node degree}
\usage{
plotTrophLevel(g, vertexLabel = FALSE, vertexSizeFactor = 5, tk = FALSE,
  modules = FALSE, ...)
}
\arguments{
\item{vertexLabel}{logical plot vertex labels}

\item{vertexSizeFactor}{numeric factor to determine the size of the label with degree}

\item{tk}{TRUE generate an interactive plot using tkplot}

\item{modules}{FALSE plot modules in the x axis, obtained with the cluster spinglass algorithm}

\item{...}{Addittional parameters to the plot function}

\item{ig}{igraph object}

\item{title}{title for the plot}
}
\value{
a plot
}
\description{
Plot ecological network organized by trophic level, with node size determined by the node degree
}
\examples{

plotTrophLevel(netData[[1]])
}
