% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{readMultiplex}
\alias{readMultiplex}
\title{Read ecological multiplex networks using different files for each layer as a 'node-aligned' multilayer network}
\usage{
readMultiplex(fileName, filePath = NULL, fhead = TRUE, skipColumn = 1)
}
\arguments{
\item{fileName}{vector of fileNames with the networks}

\item{filePath}{path of the files NULL by default}

\item{fhead}{TRUE if the files have header fields, FALSE otherwise.}

\item{skipColum}{integer, number of columns that are skiped 1 by default}
}
\value{
an igraph object if there is only one file or a list of igraph objects named after the list without extension
}
\description{
This functions uses \code{\link[=readNetwork]{readNetwork()}}
}
\examples{

# Reads a network in edge list (interaction list) format, with predators as the first column
#
fileName <- system.file("extdata", "WeddellSea_FW.csv", package = "EcoNetwork")
g <- readNetwork(fileName)

# Reads a network in adyacency matrix format, with predators as columns
#
fileName <- system.file("extdata", "BarentsBoreal_FW.csv", package = "EcoNetwork")
g <- readNetwork(fileName)

# Read a vector of files
#
\dontrun{
dn <- list.files("inst/extdata",pattern = "^.*\\\\.csv$")
netData <- readNetwork(dn,"inst/extdata")
}
}
