% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffling.R
\name{shuffle_network_ws}
\alias{shuffle_network_ws}
\title{Random Network Rewiring Without Preserving Degree Distribution}
\usage{
shuffle_network_ws(input_graph, delta = 100, directed = TRUE, weighted = TRUE)
}
\arguments{
\item{input_graph}{A square adjacency matrix (directed, weighted or unweighted) or an igraph object.}

\item{delta}{Number of rewiring attempts to perform.}

\item{directed}{Logical, whether the network is directed (default: TRUE).}

\item{weighted}{Logical, whether the network is weighted (default: TRUE).}
}
\value{
A shuffled adjacency matrix preserving the number of links but not degree distribution.
}
\description{
This function randomly rewires a directed network while preserving
the total number of links but NOT the degree distribution. This is based
on an approach described by Watts and Strogatz (1998) for small-world networks.
}
\references{
Watts, D. J., & Strogatz, S. H. (1998). Collective dynamics of 'small-world' networks.
Nature, 393(6684), 440-442. \doi{10.1038/30918}
}
