% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNetwork.r
\name{toGLVadjMat}
\alias{toGLVadjMat}
\title{From multiple interaction object 'mgraph'to GLV adjacency matrix}
\usage{
toGLVadjMat(mg, types = c("Competitive", "Mutualistic", "Trophic"),
  istrength = FALSE)
}
\arguments{
\item{mg}{multiple interaction object, class 'mgraph'}

\item{types}{vector of types that represent the layers}
}
\value{
an matrix with the signs of the interactions
}
\description{
This functions takes a 'mgraph' object and convert it to a Generalized Lotka-Volterra adjacency matrix
Position is important in fact the order is  Competitive/Negative,Mutualistic/Positive,Trophic/Antagonistic
}
\examples{

# Read a vector of files
#
\dontrun{

fpath <- system.file("extdata", package = "EcoNetwork")
dn <- list.files(fpath,pattern = "^Kefi2015.*\\\\.txt$")
netData <- readMultiplex(dn,c("Competitive","Mutualistic","Trophic"),fpath,skipColum=2)
toGLVadjMat(netData)
}
}
\seealso{
\code{\link[=readMultiplex]{readMultiplex()}}
}
